fix : filter state asset all
